feat: Rename api related to protobuf
