+++ tonic-reflection/src/server.rs
@@ -1,7 +1,7 @@
-use crate::proto::server_reflection_request::MessageRequest;
-use crate::proto::server_reflection_response::MessageResponse;
-pub use crate::proto::server_reflection_server::{ServerReflection, ServerReflectionServer};
-use crate::proto::{
+use crate::pb::server_reflection_request::MessageRequest;
+use crate::pb::server_reflection_response::MessageResponse;
+pub use crate::pb::server_reflection_server::{ServerReflection, ServerReflectionServer};
+use crate::pb::{
     ExtensionNumberResponse, FileDescriptorResponse, ListServiceResponse, ServerReflectionRequest,
     ServerReflectionResponse, ServiceResponse,
 };
@@ -109,7 +109,7 @@ impl<'b> Builder<'b> {
     /// Build a gRPC Reflection Service to be served via Tonic.
     pub fn build(mut self) -> Result<ServerReflectionServer<impl ServerReflection>, Error> {
         if self.include_reflection_service {
-            self = self.register_encoded_file_descriptor_set(crate::proto::FILE_DESCRIPTOR_SET);
+            self = self.register_encoded_file_descriptor_set(crate::pb::FILE_DESCRIPTOR_SET);
         }
 
         for encoded in &self.encoded_file_descriptor_sets {
